+++ intrinio/endpoints.py
@@ -47,6 +47,7 @@ def indices(identifier=None, query=None, type=None):
         identifier: Intrinio symbol associated with the index
         query: Search of index name or symbol
         type: Type of indices: stock_market | economic | sic
+
     Returns:
         Dataset as a Pandas DataFrame
     """
@@ -189,7 +190,7 @@ def financials_period(identifier, fiscal_year, fiscal_period='FY',
     else:
         index = pd.MultiIndex.from_tuples([(fiscal_year, fiscal_period)],
                                           names=['year', 'period'])
-                                        
+
     r = get('financials/standardized',
             identifier=identifier,
             statement=statement,
@@ -204,6 +205,32 @@ def financials_period(identifier, fiscal_year, fiscal_period='FY',
         r.index.name = 'year'
 
     return r
+
+
+def screener(conditions, order_column=None, order_direction=None,
+             primary_only=None, logic=None):
+    """
+    Find securities that meet a list of conditions.
+
+    Args:
+        conditions: Comma-separated list of conditions. Each condition
+            consists of three or four elements separated by tildes (~):
+            Data_tag~Operator~Value~Label(Optional)
+        order_column: A data tag by which to order the results
+        order_direction: Order of the results: asc (default) | desc
+        primary_only: Return only primary securities (excluding special
+            securities such as preferred shares)
+        logic: How the conditions are applied using AND by default
+
+    Returns:
+        List of tickers that meet the conditions
+    """
+    return get('securities/search',
+               conditions=conditions,
+               order_column=order_column,
+               order_direction=order_direction,
+               primary_only=primary_only,
+               logic=logic)
 
 
 def _lower_optional(x):